feat: :art: Improved separation of concerns.
Removed anything map-related from DOM and vice-versa.
