+++ src/main.ts
@@ -6,10 +6,10 @@ import {
   hideForecastLoader,
   closePopup,
   displayForecastLoader,
-  handleMapClick,
   handleMainCityClick,
+  fetchPopup,
 } from "./domManipulation";
-import { getMap, resetMap } from "./map";
+import { getMap, resetMap, bindMapClick, loadMarker } from "./map";
 
 // Only a loader for the pop-up. Ensures it only displays if results haven't been returned yet.
 displayForecastLoader("custom");
@@ -21,7 +21,7 @@ if (mainCities){
   const mainCityButtons = mainCities.getElementsByClassName("city-chip");
   // Add a click event to each of the main city chips to show their respective forecasts.
   for (let i = 0; i < mainCityButtons.length; i++) {
-    mainCityButtons[i].addEventListener("click", () => handleMainCityClick(i));
+    mainCityButtons[i].addEventListener("click", () => handleMainCityClick(i, loadMarker));
   }
 }
 
@@ -64,14 +64,12 @@ fetchWeatherForecast(
   hideForecastLoader,
 );
 
-// Set the initial view of the map to that of South Africa.
-const map = getMap();
+getMap();
+// Add a click event to the map to process when a custom or random location is selected on the map.
+bindMapClick(fetchPopup);
 
 // Display the forecast for Pretoria as default.
-handleMainCityClick(0);
-
-// Add a click event to the map to process when a custom or random location is selected on the map.
-map.on("click", handleMapClick);
+handleMainCityClick(0, loadMarker);
 
 // Add the click event to the close icon to trigger the close function.
 // Add the click event to the reset button to trigger the reset fucnction.
@@ -79,7 +77,7 @@ const closePopupButton = document.getElementById("close-popup-button");
 const resetMapButton = document.getElementById("reset-map-button");
 
 if (closePopupButton && resetMapButton) {
-  closePopupButton.addEventListener("click", () => closePopup());
+  closePopupButton.addEventListener("click", () => closePopup(loadMarker));
   resetMapButton.addEventListener("click", () => resetMap());
 }
