+++ src/map.ts
@@ -1,4 +1,5 @@
-import L from "leaflet";
+import L, { LatLng } from "leaflet";
+
 
 const markerIcon = L.icon({
   iconUrl: "https://unpkg.com/leaflet@1.9.4/dist/images/marker-icon-2x.png",
@@ -12,6 +13,8 @@ const markerIcon = L.icon({
 
 const map = L.map("map").setView([-28.921631, 25.224609], 4);
 // Set the initial marker location ot that of Pretoria as this is the default.
+
+
 const marker = L.marker([-25.73134, 28.21837], { icon: markerIcon }).addTo(map);
 
 export function getMap() {
@@ -20,7 +23,22 @@ export function getMap() {
     attribution:
       "&copy; <a href=\"http://www.openstreetmap.org/copyright\">OpenStreetMap</a>",
   }).addTo(map);
-  return map;
+
+}
+
+// Controls what happens when user clicks on the map to select own location.
+export function bindMapClick( callbackMapClick : (lat: string, long: string) => void){ 
+  map.addEventListener("click", (mapClickEvent: { latlng:  LatLng }) => {
+    const latLong = String(mapClickEvent.latlng.wrap());
+    const lat = latLong.substring(latLong.indexOf("(") + 1, latLong.indexOf(","));
+    const long = latLong.substring(
+      latLong.indexOf(",") + 2,
+      latLong.indexOf(")"),
+    );
+    loadMarker(Number(lat), Number(long));
+    callbackMapClick(lat, long);
+  });
+
 }
 
 export function loadMarker(lat: number, long: number) {
